requirements.txt
fastapi==0.108.0
uvicorn[default] ==0.25.0
sqlalchemy[asyncio] ==2.0.25
pydantic_settings==2.1.0
aiosqlite==0.19.0
pydantic[email]==2.5.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4